Add files via upload
This LM35 temperature sensor library is built for Arduino. The main function is to let the user specify the analog input pin of the Arduino to be used, and use its built-in method to read Celsius and Fahrenheit respectively. With the Kelvin temperature, they both return the temperature value of the floating point value.